refactor: added business logic

+++ src/main/java/com/example/demo/student/StudentService.java
@@ -1,6 +1,7 @@
 package com.example.demo.student;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -17,10 +18,24 @@ public class StudentService {
 	}
 
 	public void addStudent(Student student){
+		
+		Optional<Student> studentOptional = 
+		studentRepository.findStudentByEmail(student.getEmail());
+		
+		if(studentOptional.isPresent()){
+			throw new IllegalStateException("email taken");	
+		}
 		studentRepository.save(student);
 	}
 
 	public void deleteStudent(Long id){
+		boolean exists = studentRepository.existsById(id);
+		if(!exists) {
+			throw new IllegalStateException(
+				"student with id" + id + "does not exist"
+			);
+			}
 		studentRepository.deleteById(id);
-	}
+		
+}
 }
